+++ src/presentation/controllers/AuthController.ts
@@ -6,7 +6,7 @@ import { CodedError } from '../../domain/models/CodedError'
 import { TokenUseCase } from '../../application/usecases/TokenUseCase'
 import { TokenData } from '../../domain/models/TokenData'
 import { ProfileRepository } from '../../infrastructure/adapters/repositories/ProfileRepository'
-import { ProfileService } from '../../domain/services/ProfileService'
+import { ProfileService } from '../../infrastructure/adapters/services/ProfileService'
 import { ProfileUseCase } from '../../application/usecases/ProfileUseCase'
 import logger from '../../infrastructure/Logger'
 import { MongoUserRepository } from '../../infrastructure/adapters/mongoose/UserRepository'
